+++ Matrix.h
@@ -5,14 +5,14 @@
 #ifndef PROG_CONC_DIFF_MATRIX_C_H
 typedef struct {
     int size;
-    int * matrix;
+    float * matrix;
     int middle_index;
-    int max_temp_value;
+    float max_temp_value;
 } matrix_representation;
 
 
 
-int init_matrix(int, int, matrix_representation *);
+int init_matrix(int, float, matrix_representation *);
 
 void print_matrix(matrix_representation *);
